requirements.txt
Django=1.11.27
django-ckeditor=5.8.0
django-crispy-forms=1.8.1
django-formtools=2.1
django-import-export=2.0
django-js-asset=1.2.2
django-reversion=3.0.5
future=0.15.2
httplib2=0.9.2
mysql-client=0.0.1
openpyxl=3.0.2
Pillow=7.0.0
pip=19.3.1
PyMySQL=0.9.3
requests=2.22.0
six=1.10.0
urllib3=1.25.7
xadmin=0.6.1
xlwt=1.3.0
